Fix backbutton issue
Use an array to record the steps to fix the backbutton issue.
Also do some small refactors in IJarQuickPickItem.
