+++ src/exportJarFileCommand.ts
@@ -7,20 +7,11 @@ import { sendOperationError } from "vscode-extension-telemetry-wrapper";
 import { buildWorkspace } from "./build";
 import { GenerateJarExecutor } from "./exportJarSteps/GenerateJarExecutor";
 import { IExportJarStepExecutor } from "./exportJarSteps/IExportJarStepExecutor";
+import { IStepMetadata } from "./exportJarSteps/IStepMetadata";
 import { ResolveJavaProjectExecutor } from "./exportJarSteps/ResolveJavaProjectExecutor";
 import { ResolveMainMethodExecutor } from "./exportJarSteps/ResolveMainMethodExecutor";
 import { isStandardServerReady } from "./extension";
 import { INodeData } from "./java/nodeData";
-
-export interface IStepMetadata {
-    entry?: INodeData;
-    workspaceUri?: Uri;
-    isPickedWorkspace: boolean;
-    projectList?: INodeData[];
-    selectedMainMethod?: string;
-    outputPath?: string;
-    elements: string[];
-}
 
 export enum ExportJarStep {
     ResolveJavaProject = "RESOLVEJAVAPROJECT",
@@ -48,8 +39,8 @@ export async function createJarFile(node?: INodeData) {
         let step: ExportJarStep = ExportJarStep.ResolveJavaProject;
         const stepMetadata: IStepMetadata = {
             entry: node,
-            isPickedWorkspace: false,
             elements: [],
+            steps: [],
         };
         while (step !== ExportJarStep.Finish) {
             try {
