+++ src/exportJarSteps/IExportJarStepExecutor.ts
@@ -1,8 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
-import { ExportJarStep, IStepMetadata } from "../exportJarFileCommand";
+import { ExportJarStep } from "../exportJarFileCommand";
+import { IStepMetadata } from "./IStepMetadata";
 
 export interface IExportJarStepExecutor {
+    getNextStep(): ExportJarStep | undefined;
     execute(stepMetadata?: IStepMetadata): Promise<ExportJarStep>;
 }
