+++ src/exportJarSteps/ResolveJavaProjectExecutor.ts
@@ -3,18 +3,23 @@
 
 import * as _ from "lodash";
 import { Disposable, Uri, workspace } from "vscode";
-import { ExportJarStep, IStepMetadata } from "../exportJarFileCommand";
+import { ExportJarStep } from "../exportJarFileCommand";
 import { Jdtls } from "../java/jdtls";
 import { INodeData } from "../java/nodeData";
 import { WorkspaceNode } from "../views/workspaceNode";
 import { IExportJarStepExecutor } from "./IExportJarStepExecutor";
+import { IStepMetadata } from "./IStepMetadata";
 import { createPickBox, IJarQuickPickItem } from "./utility";
 
 export class ResolveJavaProjectExecutor implements IExportJarStepExecutor {
 
+    public getNextStep(): ExportJarStep {
+        return ExportJarStep.ResolveMainMethod;
+    }
+
     public async execute(stepMetadata: IStepMetadata): Promise<ExportJarStep> {
         await this.resolveJavaProject(stepMetadata, stepMetadata.entry);
-        return ExportJarStep.ResolveMainMethod;
+        return this.getNextStep();
     }
 
     private async resolveJavaProject(stepMetadata: IStepMetadata, node?: INodeData): Promise<void> {
@@ -33,28 +38,27 @@ export class ResolveJavaProjectExecutor implements IExportJarStepExecutor {
         const pickItems: IJarQuickPickItem[] = [];
         const projectMap: Map<string, INodeData[]> = new Map<string, INodeData[]>();
         for (const folder of folders) {
-            const uriString: string = folder.uri.toString();
-            const projects: INodeData[] = await Jdtls.getProjects(uriString);
+            const projects: INodeData[] = await Jdtls.getProjects(folder.uri.toString());
             if (!_.isEmpty(projects)) {
                 pickItems.push({
                     label: folder.name,
                     description: folder.uri.fsPath,
-                    uri: uriString,
+                    uri: folder.uri,
                 });
-                projectMap.set(uriString, projects);
+                projectMap.set(folder.uri.toString(), projects);
             }
         }
         if (_.isEmpty(pickItems)) {
             throw new Error("No java project found. Please make sure your Java project exists in the workspace.");
         }
-        stepMetadata.isPickedWorkspace = true;
         const disposables: Disposable[] = [];
         await new Promise((resolve, reject) => {
             const pickBox = createPickBox("Export Jar : Determine project", "Select the project", pickItems, false);
             disposables.push(
                 pickBox.onDidAccept(() => {
-                    stepMetadata.workspaceUri = Uri.parse(pickBox.selectedItems[0].uri);
-                    stepMetadata.projectList = projectMap.get(pickBox.selectedItems[0].uri);
+                    stepMetadata.workspaceUri = pickBox.selectedItems[0].uri;
+                    stepMetadata.projectList = projectMap.get(pickBox.selectedItems[0].uri.toString());
+                    stepMetadata.steps.push(ExportJarStep.ResolveJavaProject);
                     return resolve();
                 }),
                 pickBox.onDidHide(() => {
