+++ src/exportJarSteps/ResolveMainMethodExecutor.ts
@@ -2,9 +2,10 @@
 // Licensed under the MIT license.
 
 import { Disposable, ProgressLocation, QuickInputButtons, window } from "vscode";
-import { ExportJarStep, IStepMetadata } from "../exportJarFileCommand";
+import { ExportJarStep } from "../exportJarFileCommand";
 import { Jdtls } from "../java/jdtls";
 import { IExportJarStepExecutor } from "./IExportJarStepExecutor";
+import { IStepMetadata } from "./IStepMetadata";
 import { createPickBox, IJarQuickPickItem } from "./utility";
 
 export class ResolveMainMethodExecutor implements IExportJarStepExecutor {
@@ -13,11 +14,15 @@ export class ResolveMainMethodExecutor implements IExportJarStepExecutor {
         return data.name.substring(data.name.lastIndexOf(".") + 1);
     }
 
+    public getNextStep(): ExportJarStep {
+        return ExportJarStep.GenerateJar;
+    }
+
     public async execute(stepMetadata: IStepMetadata): Promise<ExportJarStep> {
         if (await this.resolveMainMethod(stepMetadata)) {
-            return ExportJarStep.GenerateJar;
+            return this.getNextStep();
         }
-        return ExportJarStep.ResolveJavaProject;
+        return stepMetadata.steps.pop();
     }
 
     private async resolveMainMethod(stepMetadata: IStepMetadata): Promise<boolean> {
@@ -54,7 +59,7 @@ export class ResolveMainMethodExecutor implements IExportJarStepExecutor {
         try {
             result = await new Promise<boolean>(async (resolve, reject) => {
                 const pickBox = createPickBox("Export Jar : Determine main class", "Select the main class",
-                    pickItems, stepMetadata.isPickedWorkspace);
+                    pickItems, stepMetadata.steps.length > 0);
                 disposables.push(
                     pickBox.onDidTriggerButton((item) => {
                         if (item === QuickInputButtons.Back) {
@@ -63,6 +68,7 @@ export class ResolveMainMethodExecutor implements IExportJarStepExecutor {
                     }),
                     pickBox.onDidAccept(() => {
                         stepMetadata.selectedMainMethod = pickBox.selectedItems[0].description;
+                        stepMetadata.steps.push(ExportJarStep.ResolveMainMethod);
                         return resolve(true);
                     }),
                     pickBox.onDidHide(() => {
