+++ src/exportJarSteps/utility.ts
@@ -1,10 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
-import { QuickInputButtons, QuickPick, QuickPickItem, window } from "vscode";
+import { QuickInputButtons, QuickPick, QuickPickItem, Uri, window } from "vscode";
 
 export interface IJarQuickPickItem extends QuickPickItem {
-    uri?: string;
+    uri?: Uri;
+    path?: string;
     type?: string;
 }
